requirements.txt
aiohttp>=3.9.0
aiogram>=3.0.0
clickhouse-connect>=0.7.0
psutil>=5.9.0
python-dotenv>=1.0.0